+++ src/Graphify/Syntax/TypeDeclarationSyntaxExtensions.ToSubject.cs
@@ -48,7 +48,7 @@
             SemanticModel model = compilation.GetSemanticModel(syntax.SyntaxTree);
             ISymbol symbol = model.GetDeclaredSymbol(syntax, cancellationToken: cancellationToken);
 
-            if (!(symbol is INamedTypeSymbol type && type.HasGraphify(out byte depth)))
+            if (!(symbol is INamedTypeSymbol type && type.HasGraphify(out byte depth) && type.HasSupportedAccessibility()))
             {
                 return default;
             }